% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relRisk.R
\name{relRisk}
\alias{relRisk}
\title{Calculate Relative Risk for Individual Alleles and Genotypes in BIGDAWG-formatted Case-Control Datasets}
\usage{
relRisk(dataset, return = TRUE)
}
\arguments{
\item{dataset}{The name of a BIGDAWG formatted case-control genotype dataset. Either a tab-delimited file or a data frame can be specified.}

\item{return}{A logical identifying if the list object should be returned (return=TRUE), or if pairs of tab-delimited text files of results (one for alleles and one for genotypes) should be written to the working directory for each locus.}
}
\value{
A list object of two lists ("alleles" and "genotypes"), each of which contains a list of nine-column data frames containing results for each unique allele or genotype (in rows) at each locus. Column headers in each dataframe are, *Locus*, *Variant*, *Cases*, *Controls*, *RelativeRisk*, *CI.low*, *CI.high*, *p.value*, and *Significant*.
}
\description{
This function returns a list object containing relative risk, confidence interval and p-value data for the individual alleles and individual genotypes at each locus in a BIGDAWG-formatted case-control genotype data frame or file.
}
\examples{
# Analyzing the BIGDAWG::HLA_data dataset 
library(BIGDAWG)
rr <- relRisk(HLA_data[,1:4])
}
\references{
\href{https://cran.r-project.org/web/packages/BIGDAWG/vignettes/BIGDAWG.html#input-data}{BIGDAWG Data Format}
}
\keyword{allele}
\keyword{genotype}
\keyword{relative}
\keyword{risk}
